Add ability to do incremental backups and also...
A lot of various changes to make the whole thing work better.
These include refactoring the full backup function out so that
the incremental backup can use it when making an 'incremental'
back up a volume it wasn't in the previous backup.

Also, standardize the date format by using a global variable
for it. (Actually, a constant, but Python doesn't have real
constant constants).